Build system overhaul (#47)
* doc: update spanish guide

* feat: desktop shortcut for linux

* doc: readme for desktop shortcuts

* fix: categories

* fix: wrong name

* fix: remove useless and forgotten about define

* fix: assume we're installing in /opt, add terminal=true

* feat: overhaul build system

* fix: lower required boost version

* fix: include order dependency

* fix: macOS build?

* fix: try build static for macOS

* fix: macOS build?

* chore: macOS deployment target should be 10.10

* fix: CXXFlags don't work try SYSROOT

* fix: findpcap still required for windows

* chore: increase minimum sdk to 10.11

so I can get rid of the _objc_loadClassref linker flag again, don't trust it

* chore: get rid of questionable linker flag

* fix: with ARC working on 10.11 no manual management is needed

* doc: extended static compile instructions

* doc: add macOS instructions

* chore: fix warning

* chore: went over clang tidy suggestions

* fix: tests have a different name now

+++ Sources/WifiInterfaceLinuxBSD.cpp
@@ -1,6 +1,6 @@
 /* Copyright (c) 2021 [Rick de Bondt] - WifiInterfaceBSD.cpp */
 
-#include "../Includes/WifiInterfaceLinuxBSD.h"
+#include "WifiInterfaceLinuxBSD.h"
 
 #include <cerrno>
 #include <chrono>
@@ -8,8 +8,8 @@
 #include <ifaddrs.h>
 #include <net/if.h>
 
-#include "../Includes/Logger.h"
-#include "../Includes/NetConversionFunctions.h"
+#include "Logger.h"
+#include "NetConversionFunctions.h"
 
 #ifdef __linux__
 #include <netpacket/packet.h>
@@ -150,7 +150,7 @@ static int FamilyHandler(nl_msg* aMessage, void* aArgument)
                     static_cast<size_t>(nla_len(lMulticastGroupIndices.at(CTRL_ATTR_MCAST_GRP_NAME)) - 1)};
 
                 if (lCurrentGroup == lGroup->group) {
-                    lGroup->id = nla_get_u32(lMulticastGroupIndices.at(CTRL_ATTR_MCAST_GRP_ID));
+                    lGroup->id = static_cast<int>(nla_get_u32(lMulticastGroupIndices.at(CTRL_ATTR_MCAST_GRP_ID)));
                     // No other solution than calling break here sadly
                     break;
                 }
@@ -312,7 +312,7 @@ static int DumpResults(nl_msg* aMessage, void* aArgument)
 
                     // Grab the frequency
                     if (lBSS.at(NL80211_BSS_FREQUENCY) != nullptr) {
-                        lInformation.frequency = nla_get_u32(lBSS.at(NL80211_BSS_FREQUENCY));
+                        lInformation.frequency = static_cast<int>(nla_get_u32(lBSS.at(NL80211_BSS_FREQUENCY)));
                     }
 
                     // Using the Capability field, index 0, field ESS, if this is 0, the network is an adhoc
@@ -562,7 +562,7 @@ bool WifiInterface::Connect(const IWifiInterface::WifiInformation& aConnection)
         // Interface, what ssid to connect to, what bssid to connect to and frequency
         nla_put_u32(
             lMessage, NL80211_ATTR_IFINDEX, mNetworkAdapterIndex);  // Add message attribute, which interface to use.
-        nla_put(lMessage, NL80211_ATTR_SSID, aConnection.ssid.length(), aConnection.ssid.data());
+        nla_put(lMessage, NL80211_ATTR_SSID, static_cast<int>(aConnection.ssid.length()), aConnection.ssid.data());
         nla_put_flag(lMessage, NL80211_ATTR_FREQ_FIXED);
         nla_put_u32(lMessage,
                     NL80211_ATTR_WIPHY_FREQ,
